Refactor y moví algunas cosas
