+++ app/src/main/res/layout/fragment_list_ingredients.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.ListIngredients" >
+    tools:context=".fragments.main.ListIngredients" >
 
     <TextView
         android:id="@+id/textView"
